# `filtering_columns_first` for reading Parquet/ORC in cuDF/Dask cuDF