--- a/python/cudf/cudf/utils/ioutils.py
+++ b/python/cudf/cudf/utils/ioutils.py
@@ -6,6 +6,7 @@ import urllib
 from io import BufferedWriter, BytesIO, IOBase, TextIOWrapper
 from threading import Thread
 
+import cudf
 import fsspec
 import fsspec.implementations.local
 import numpy as np
@@ -1539,6 +1540,19 @@ def _filters_to_query(filters):
         query_string += ")"
 
     return query_string, local_dict
+
+
+def _make_empty_df(schema, columns=None):
+    col_names = schema.names if columns is None else columns
+    return cudf.DataFrame(
+        {
+            col_name: cudf.core.column.column_empty(
+                row_count=0,
+                dtype=schema.field(col_name).type.to_pandas_dtype(),
+            )
+            for col_name in col_names
+        }
+    )
 
 
 def _ensure_filesystem(passed_filesystem, path):