# Summary

This PR adds an optional `filtering_columns_first` parameter to `cudf.read_parquet`,  `cudf.read_orc`, `dask_cudf.read_parquet`, and `dask_cudf.read_orc`. When set `filtering_columns_first=True`, specified `filters` are used to first read in columns relevant to filtering and only read in the remaining columns if the filtered filtering columns isn't empty.

This PR includes:
- `filtering_columns_first=True` optimized case for `cudf.read_parquet` and `cudf.read_orc`
- `filtering_columns_first=True` for `dask_cudf.read_parquet` and `dask_cudf.read_orc` (which simply passes through to `cudf`)
- Tests and docs for `cudf.read_parquet`,  `cudf.read_orc`, `dask_cudf.read_parquet`, and `dask_cudf.read_orc`

# Previously...

Previously, I implemented `filtering_columns_first` directly in Dask cuDF with https://github.com/rapidsai/cudf/pull/9501. This PR takes a different approach of implementing `filtering_columns_first` directly in cuDF and then passing the parameters in from the Dask `read_*` functions.

There are several advantages of this approach:
- Simpler implementation 
- This PR is able to add `filtering_columns_first` for both `read_orc` and `read_parquet.
- Able to use this in cuDF as well as Dask cuDF
- Lazy processing of filtering columns

The disadvantage is several instances of overhead:
- Extra scheduling because of extra partitions (https://github.com/rapidsai/cudf/pull/9501 would cull partitions before reading in remaining columns)
- Reading metadata in each partition to return an empty dataframe
- Concatenating together dataframes from the greater number of partitions when `compute`ing

# Example and Benchmark

## ORC

Without any `filters`:
```
df = dask_cudf.read_orc(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.orc"
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 27 s
```

With `filters` used for metadata statistics-based partition pruning:
```
df = dask_cudf.read_orc(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.orc"
    filters=[("ws_item_sk", "==", 7947), ("ws_bill_customer_sk", "==", 854575)]
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 9.04 s
```

With the old implementation of `filtering_columns_first=True` in https://github.com/rapidsai/cudf/pull/9501:
```
df = dask_cudf.read_orc(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.orc"
    filters=[("ws_item_sk", "==", 7947), ("ws_bill_customer_sk", "==", 854575)],
    filtering_columns_first=True
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 1.73 s
```

With the new implementation of `filtering_columns_first=True`:
```
df = dask_cudf.read_orc(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.orc"
    filters=[("ws_item_sk", "==", 7947), ("ws_bill_customer_sk", "==", 854575)],
    filtering_columns_first=True
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 2.28 s
# 4x speedup
```

## Parquet

Without any `filters`:
```
df = dask_cudf.read_parquet(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.parquet"
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 13 s
```

With `filters` used for metadata statistics-based partition pruning:
```
df = dask_cudf.read_parquet(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.parquet"
    filters=[("ws_item_sk", "==", 7947), ("ws_bill_customer_sk", "==", 854575)]
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 13 s
```

With the new implementation of `filtering_columns_first=True`:
```
df = dask_cudf.read_parquet(
    "datasets/gpu-bdb/sf100/orc-split-row-group/web_sales/*.parquet"
    filters=[("ws_item_sk", "==", 7947), ("ws_bill_customer_sk", "==", 854575)],
    filtering_columns_first=True
).query("ws_item_sk == 7947 and ws_bill_customer_sk == 854575").compute()
# Wall time: 5.6 s
# 2.3x speedup
```

# Performance

Note that the default still has `filters` specified and the `filters` is used to prune partitions based on stripe/row group statistics in Parquet/ORC metadata. The performance improvement comes from reading filtering columns first and using this to eliminate
unnecessary I/O (on the partitions that get emptied by applying `filters` query to the filtering columns).

## ORC (4x speedup)

![Better scalability for many-column ORC datasets](https://i.imgur.com/OcVPgpq.png)

## Parquet (2.3x speedup)

![Better scalability for many-column Parquet datasets](https://i.imgur.com/oxf6WBN.png)

# FAQ

### Why are the ORC/Parquet and cuDF/Dask cuDF implementations in the same PR?

Few reasons:
1. They share a couple common helper functions in `ioutils.py`.
2. The ORC implementation is simpler but would if it were in its own PR first, it would be deprioritized and that would just delay the Parquet implementation.
3. This can be prioritized higher because of the greater value add for Dask cuDF users.